test: split React component tests into multiple files
